requirements.txt
# Core Backend Framework for Valuation Application
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# MongoDB and Database
motor==3.6.0
pymongo>=4.9,<4.10
dnspython==2.7.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
boto3==1.35.77
botocore==1.35.77
cachetools==5.5.0

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP and API
httpx==0.28.1
requests==2.32.3

# Data Processing
pandas==2.1.4
numpy==1.26.4

# File Processing
python-docx==1.1.2
pypdf2==3.0.1
pillow==10.4.0
openpyxl==3.1.5

# Email and Communications
emails==0.6
jinja2==3.1.4

# Logging and Monitoring
structlog==24.4.0
colorama==0.4.6

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1

# Date and Time
python-dateutil==2.9.0.post0
pytz==2024.2

# Utility Libraries
click==8.1.7
rich==13.9.4
typer==0.15.1

# Production Server
gunicorn==23.0.0

# Additional Security
cryptography==43.0.3
bcrypt==4.2.1
email-validator==2.2.0

# File upload and processing
aiofiles==24.1.0

# Image processing (for property photos)
opencv-python-headless==4.10.0.84

# Excel/CSV processing
xlsxwriter==3.2.0

# Background tasks (for report generation)
celery==5.4.0
redis==5.2.0

# Environment-specific packages
python-magic==0.4.27

# Performance monitoring
psutil==6.1.0

# Development utilities
watchdog==6.0.0

# Web scraping (for property data)
beautifulsoup4==4.12.3
mypy==1.13.0

# Documentation
sphinx==8.1.3

# PDF Generation
weasyprint==61.2

# Jupyter notebooks (for data analysis)
jupyter==1.1.1
ipykernel==6.29.5